Terminal

See merge request cs5031-p3-group07/cs5031p3code!3

+++ EstateAgencySystemTerminal/src/main/java/stacs/estate/cs5031p3code/menu/UserMenu.java
@@ -5,6 +5,12 @@ import stacs.estate.cs5031p3code.client.TerminalClient;
 
 import java.util.Scanner;
 
+/**
+ * An implementation of the {@link Menu}.
+ * This is the user menu which allows user to access user data.
+ *
+ * @author 190005675
+ */
 public class UserMenu implements Menu {
     private final Scanner scanner;
     private final TerminalClient client;
@@ -56,6 +62,10 @@ public class UserMenu implements Menu {
                     case "2" -> {
                         System.out.println("Please enter the id of the user to be updated: ");
                         String userId = scanner.nextLine();
+                        if (userId.equals("")) {
+                            response = "ID cannot be empty!";
+                            break;
+                        }
                         System.out.println("Please enter new details below (empty string will be ignored)");
                         System.out.println("Email: ");
                         String email = scanner.nextLine();
@@ -72,6 +82,10 @@ public class UserMenu implements Menu {
                     case "3" -> {
                         System.out.println("Please enter the id of the user to be deleted: ");
                         String userId = scanner.nextLine();
+                        if (userId.equals("")) {
+                            response = "ID cannot be empty!";
+                            break;
+                        }
                         response = client.deleteUserById(userKey, userId);
                     }
                     case "4" -> {
@@ -91,4 +105,3 @@ public class UserMenu implements Menu {
         }
     }
 }
-
